Add updating of main indicator on change
